Only add existing files to the ruby requires
This is a guard to not add to the list files that don't exist.
When sm is installed for the root user, those files are

    /opt/sm/core/sm/shell/ruby/core/functions.rb
    /opt/sm/core/sm/shell/ruby/core/initialize.rb

They look like core sm files: this fix is not the definitive one.
Are these files located elsewhere?
Should this block of code be entirely removed?